Harvest API example
This file contains an example of calling the Harvest API. The access token and account ID have been masked. There's no visualization in the report. This file was created solely to test a query and a function that calls the Harvest API. Documentation for the API is available at https://help.getharvest.com/api-v2/